[FIX] detail의 버튼 분리, id 위치 수정
수덩
