+++ config/customer/customer.js
@@ -137,24 +137,24 @@ export const COLUMN_NAME = {
     "처리 상태"
   ],
   VIEW_COMPLAINT: [
-    "민원 종류",
     "민원 번호",
+    "민원 종류",
     "제목",
     "등록 날짜",
     "처리된 날짜",
     "처리상태"
   ],
   INSURANCE_LIST: [
+    "보험 상품 번호",
     "보험 상품 이름",
     "보험 종류",
-    "보험 상품 번호",
     "연령대",
     "월 보험료(원)"
   ],
   LOAN_LIST: [
+    "대출 상품 번호",
     "대출 상품 이름",
     "대출 상품 종류",
-    "대출 상품 번호",
     "이자율",
     "대출가능 최대 금액(원)"
   ]
