+++ js/utils/uiUtils/buttonManager/employee/compensationPlanning/detail.js
@@ -1,0 +1,22 @@
+import {initialButtons} from "../../../common/buttonUtils.js";
+import {BUTTON} from "../../../../../../config/employee/compensationPlanning/compensationPlanning.js";
+
+const evaluate = () => {
+  alert("평가-보상기획");
+}
+
+const cancel = () => {
+  window.history.back();
+}
+
+const compensationPlanningTaskMapper = {
+  EVALUATE_PARTNERCOMPANY: {
+    EVALUATE: evaluate,
+    CANCEL: cancel
+  }
+}
+
+export const renderButtons = () => {
+  const type = sessionStorage.getItem("currentType");
+  initialButtons(BUTTON.TASK.COMPENSATIONPLANNING[type], compensationPlanningTaskMapper[type]);
+}
