+++ js/utils/uiUtils/messageManager/customer/message.js
@@ -5,11 +5,9 @@ export const renderGreeting = async () => {
   const container = document.querySelector(".container");
   const id = sessionStorage.getItem("id");
   const getCustomer = await fetchGetCustomerById(id);
-  const greetingConstant = MESSAGES.GREETING;
   const greeting = document.createElement("div");
   greeting.className = "greeting";
-
-  greeting.textContent = `${getCustomer.name} ${greetingConstant} `;
+  greeting.textContent = `${getCustomer.name} ${MESSAGES.GREETING} `;
 
   const headerLine = document.querySelector(".header-line");
   container.insertBefore(greeting, headerLine.nextSibling);
