+++ js/utils/uiUtils/tableRenderer/customer/detail.js
@@ -3,8 +3,7 @@ import {
   fetchGetInsuranceByProductId,
   fetchGetLoanByProductId
 } from "../../../apiUtils/apiDocumentation/customer/customer.js";
-import {BUTTON} from "../../../../../config/customer/customer.js";
-import {initialButtons} from "../../common/buttonUtils.js";
+import {renderButtons} from "../../buttonManager/customer/detail.js";
 
 const insuranceTypeStr = {
   Disease: "질병",
@@ -141,20 +140,17 @@ const context = {
   MANAGEMENT_CONTRACT: {
     needCustomerId: true,
     detailGetter: contractDetail,
-    fetchGetById: fetchGetContractById,
-    buttons: BUTTON.TASK.CUSTOMER.MANAGEMENT_CONTRACT
+    fetchGetById: fetchGetContractById
   },
   INSURANCE_LIST: {
     needCustomerId: false,
     detailGetter: insuranceDetail,
-    fetchGetById: fetchGetInsuranceByProductId,
-    buttons: BUTTON.TASK.CUSTOMER.INSURANCE_LIST
+    fetchGetById: fetchGetInsuranceByProductId
   },
   LOAN_LIST: {
     needCustomerId: false,
     detailGetter: loanDetail,
-    fetchGetById: fetchGetLoanByProductId,
-    buttons: BUTTON.TASK.CUSTOMER.LOAN_LIST
+    fetchGetById: fetchGetLoanByProductId
   }
 }
 
@@ -216,68 +212,4 @@ const renderDetailsTable = (data) => {
     }
     detailsTable.querySelector("tbody").appendChild(row);
   });
-}
-
-const applyRecontract = () => {
-  alert("재가입 신청");
-}
-
-const applyRevival = () => {
-  alert("부활 신청");
-}
-
-const applyTermination = () => {
-  alert("해지 신청");
-}
-
-const applyEndorsement = () => {
-  alert("배서 신청");
-}
-
-const payInsuranceFee = () => {
-  alert("보험금 납입");
-}
-
-const receiveInsurance = () => {
-  alert("보험금 신청");
-}
-
-const askInsuranceCounsel = () => {
-  alert("상담 신청");
-}
-
-const buyInsurance = () => {
-  alert("보험 구매");
-}
-
-const loan = () => {
-  alert("대출 신청");
-}
-
-const cancel = () => {
-  window.history.back();
-}
-
-const customerTaskMapper = {
-  MANAGEMENT_CONTRACT: {
-    RECONTRACT: applyRecontract,
-    REVIVAL: applyRevival,
-    TERMINATION: applyTermination,
-    ENDORSEMENT: applyEndorsement,
-    PAY_INSURANCE_FEE: payInsuranceFee,
-    RECEIVE_INSURANCE: receiveInsurance
-  },
-  INSURANCE_LIST: {
-    COUNSEL: askInsuranceCounsel,
-    BUY_INSURANCE: buyInsurance
-  },
-  LOAN_LIST: {
-    APPLY: loan,
-    CANCEL: cancel
-  }
-}
-
-const renderButtons = () => {
-  const type = sessionStorage.getItem("currentType");
-  initialButtons(BUTTON.TASK.CUSTOMER[type], customerTaskMapper[type]);
 }
